+++ src/components/ListItem/index.stories.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { action } from "@storybook/addon-actions";
 import { boolean, select, text } from "@storybook/addon-knobs";
 
-import { Typography } from "../..";
+import Typography from "../Typography";
 import { Icons } from "../../types/Icon";
 import { getAllComposedDataCy } from "../../utils";
 import { getDocumentationPage, StoriesWrapper } from "../../utils/stories";
